+++ app/Self/component/cognition/Broads.jsx
@@ -7,7 +7,7 @@ import {
 } from 'react-router-dom';
 import {connect} from "react-redux";
 import NailShared from '../../../Component/Nails/NailShared.jsx';
-import Unit from '../../../Component/Unit.jsx';
+import Unit from '../../../Unit/Unit.jsx';
 
 class Broads extends React.Component {
   constructor(props){
