+++ app/Self/component/cognition/Collaterals.jsx
@@ -7,7 +7,7 @@ import {
 } from 'react-router-dom';
 import {connect} from "react-redux";
 import NailShared from '../../../Component/Nails/NailShared.jsx';
-import Unit from '../../../Component/Unit.jsx';
+import Unit from '../../../Unit/Unit.jsx';
 
 class Tracks extends React.Component {
   constructor(props){
