+++ app/Self/component/cognition/Shared.jsx
@@ -7,7 +7,7 @@ import {
 import {connect} from "react-redux";
 import Threads from './Threads.jsx';
 import TitleShared from './Titles/TitleShared.jsx';
-import Unit from '../../../Component/Unit.jsx';
+import Unit from '../../../Unit/Unit.jsx';
 import NailShared from '../../../Component/Nails/NailShared.jsx';
 //ModalBox used some unstable method, considering updating some day.
 import ModalBox from '../../../Component/ModalBox.jsx';
