+++ app/Unit/Mark/ViewerBlock/ViewerBlock.jsx
@@ -3,13 +3,10 @@ import {connect} from "react-redux";
 import classnames from 'classnames';
 import styles from "./styles.module.css";
 import PanelJump from '../PanelJump.jsx';
-import AuthorInspired from '../AuthorInspired.jsx';
+import ViewerBulb from '../ViewerBulb.jsx';
 import MarkDialogue from '../MarkDialogue.jsx';
-import SvgPropic from '../../../Svg/SvgPropic.jsx';
-import DraftDisplay from '../../../Draft/DraftDisplay.jsx';
-import {
-  setUnitInspired
-} from "../../../../redux/actions/general.js";
+import SvgPropic from '../../../Component/Svg/SvgPropic.jsx';
+import DraftDisplay from '../../../Component/Draft/DraftDisplay.jsx';
 
 const styleMiddle = {
   boxPanelInteraction: {
@@ -18,6 +15,10 @@ const styleMiddle = {
     position: 'relative',
     boxSizing: 'border-box',
   },
+  boxMessage: {
+    boxSizing: 'border-box',
+    padding: '1rem 0.7rem 0'
+  },
   spanInteractions: {
     fontSize: '1.4rem',
     letterSpacing: '0.18rem',
@@ -25,53 +26,56 @@ const styleMiddle = {
     fontWeight: '400',
     cursor: 'pointer'
   },
-  fontInteractions: {
-    fontSize: '1.4rem',
-    letterSpacing: '0.18rem',
-    lineHeight: '1.9rem',
-    fontWeight: '300',
+  textMessage: {
+    fontSize: '1.2rem',
+    letterSpacing: '0.1rem',
+    fontStyle: 'italic',
+    color: '#adadad',
+    cursor: 'default'
   }
 }
 
-class AuthorBlock extends React.Component {
+class ViewerBlock extends React.Component {
   constructor(props){
     super(props);
     this.state = {
-      dialogue: false
+      dialogue: false,
+      message: false
     };
     this.boxContent = React.createRef();
-    this._set_stateDefault = ()=>{this.setState({dialogue: false})};
+    this._set_stateDefault = ()=>{this.setState({dialogue: false, message: false})};
+    this._set_BlockMessage = this._set_BlockMessage.bind(this);
     this._handleWheel_boxContent = (event)=>{event.stopPropagation();};
     this._handleClick_openDialogue = this._handleClick_openDialogue.bind(this);
     this.style = {
-      Com_AuthorBlock_: {
+      Com_ViewerBlock_: {
         width: '100%',
         height: '100%',
         boxSizing: 'border-box',
         overflowY: 'auto'
       },
-      Com_AuthorBlock_content_: {
+      Com_ViewerBlock_content_: {
         display: 'flex',
         flexDirection: 'column',
         width: '100%',
         marginBottom:'27px'
       },
-      Com_AuthorBlock_panel_: {
+      Com_ViewerBlock_panel_: {
         width: '100%',
         height: '2.1rem',
         boxSizing: 'border-box',
         margin: '1.2rem 0px 1.8rem',
       },
-      Com_AuthorBlock_credits_: {
+      Com_ViewerBlock_credits_: {
         width: '100%',
         height: '2.6rem',
         boxSizing: 'border-box',
         marginTop: '2%',
       },
-      Com_AuthorBlock_fold_:{
+      Com_ViewerBlock_fold_:{
         display: 'none'
       },
-      Com_AuthorBlock_fold_dialogue: {
+      Com_ViewerBlock_fold_dialogue: {
         width: '100%',
         height: '100%',
         position: 'absolute',
@@ -88,6 +92,10 @@ class AuthorBlock extends React.Component {
     event.preventDefault();
     event.stopPropagation();
     this.setState((prevState, props)=>{return this.state.dialogue?{dialogue: false}: {dialogue: true}})
+  }
+
+  _set_BlockMessage(message){
+    this.setState({message: message});
   }
 
   componentDidUpdate(prevProps, prevState, snapshot){
@@ -115,10 +123,10 @@ class AuthorBlock extends React.Component {
 
     return(
       <div
-        style={this.style.Com_AuthorBlock_}>
+        style={this.style.Com_ViewerBlock_}>
         <div
           ref={this.boxContent}
-          style={Object.assign({}, this.style.Com_AuthorBlock_content_)}>
+          style={Object.assign({}, this.style.Com_ViewerBlock_content_)}>
           <div
             style={{
               width: '100%',
@@ -146,7 +154,7 @@ class AuthorBlock extends React.Component {
             }
           )}>
           <div
-            style={Object.assign({},this.style.Com_AuthorBlock_panel_)}>
+            style={Object.assign({},this.style.Com_ViewerBlock_panel_)}>
             <div
               style={Object.assign(
                 {},
@@ -163,15 +171,22 @@ class AuthorBlock extends React.Component {
               style={Object.assign(
                 {},
                 styleMiddle.boxPanelInteraction,
-                styleMiddle.fontInteractions,
-                {marginLeft: '58%',float: 'left'}
+                {marginLeft: '69%',float: 'left'}
               )}>
-              <AuthorInspired
-                markKey={this.props.markKey}/>
+              <ViewerBulb
+                markKey={this.props.markKey}
+                _set_BlockMessage={this._set_BlockMessage}/>
             </div>
           </div>
+          {
+            this.state.message &&
+            <div
+              style={styleMiddle.boxMessage}>
+              <span style={styleMiddle.textMessage}>{this.state.message}</span>
+            </div>
+          }
           <div
-            style={Object.assign({}, this.style.Com_AuthorBlock_credits_)}>
+            style={Object.assign({}, this.style.Com_ViewerBlock_credits_)}>
             <span  style={{display:'inline-block', width: "24%", height: '99%', position: 'relative'}}><SvgPropic/></span>
             <span  style={{display:'inline-block', width: "24%", height: '99%', position: 'relative'}}><SvgPropic/></span>
           </div>
@@ -180,11 +195,11 @@ class AuthorBlock extends React.Component {
           </div>
         </div>
         <div
-          style={this.style.Com_AuthorBlock_fold_}>
+          style={this.style.Com_ViewerBlock_fold_}>
           {
             this.state.dialogue &&
             <div
-              style={this.style.Com_AuthorBlock_fold_dialogue}>
+              style={this.style.Com_ViewerBlock_fold_dialogue}>
               <MarkDialogue
                 markKey={this.props.markKey}/>
             </div>
@@ -212,4 +227,4 @@ const mapDispatchToProps = (dispatch)=>{
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(AuthorBlock);
+)(ViewerBlock);
