+++ app/Unit/UnitViewSummary.jsx
@@ -7,9 +7,9 @@ import {connect} from "react-redux";
 import classnames from 'classnames';
 import MarksArticle from './MarksArticle.jsx';
 import {NodesExtensible} from './NodesDisplay/NodesDisplay.jsx';
-import DateConverter from '../DateConverter.jsx';
-import SvgCreate from '../Svg/SvgCreate.jsx'
-import {AccountPlate} from '../AccountPlate.jsx';
+import DateConverter from '../Component/DateConverter.jsx';
+import SvgCreate from '../Component/Svg/SvgCreate.jsx'
+import {AccountPlate} from '../Component/AccountPlate.jsx';
 
 const styleMiddle = {
