+++ app/Unit/ViewerPanel.jsx
@@ -4,7 +4,6 @@ import {
   withRouter
 } from 'react-router-dom';
 import {connect} from "react-redux";
-import cxBind from 'classnames/bind';
 
 class ViewerPanel extends React.Component {
   constructor(props){
