+++ app/Within/component/LtdUnits.jsx
@@ -6,7 +6,7 @@ import {
 } from 'react-router-dom';
 import {connect} from "react-redux";
 import cxBind from 'classnames/bind';
-import Unit from '../../Component/Unit.jsx';
+import Unit from '../../Unit/Unit.jsx';
 import NailScape from '../../Component/Nails/NailScape.jsx';
 import {
   handleNounsList,
