+++ app/Within/component/RelationRes.jsx
@@ -6,7 +6,7 @@ import {
 } from 'react-router-dom';
 import {connect} from "react-redux";
 import cxBind from 'classnames/bind';
-import Unit from '../../Component/Unit.jsx';
+import Unit from '../../Unit/Unit.jsx';
 import SvgPropic from '../../Component/Svg/SvgPropic.jsx';
 import DraftDisplay from '../../Component/Draft/DraftDisplay.jsx';
