+++ app/config/styleParams.js
